requirements.txt
fal-client>=0.7.0
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
